issue 463: allow to select any device from list
